separate rates and markets
