+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "zelcore-rates-api-v2",
+  "name": "rates-api",
   "version": "2.0.0",
   "description": "A service that supplies market info for ZelCore assets and conversions to other currencies",
   "main": "index.js",
