+++ src/routes.js
@@ -1,5 +1,6 @@
 const apicache = require('apicache');
 const rate = require('./controllers/rate');
+const market = require('./controllers/market');
 
 const cache = apicache.middleware;
 
@@ -8,4 +9,5 @@ module.exports = (app) => {
     res.redirect('/rates');
   });
   app.get('/rates', cache('3 minutes'), rate.list);
+  app.get('/markets', cache('3 minutes'), market.list);
 };
